move structToURLValues function into an internal package called options and also rename it to Encode
